+++ http/licenses/licenses.go
@@ -1,0 +1,8 @@
+package licenses
+
+import _ "embed"
+
+// Text contains the license information for rqlite and its dependencies.
+//
+//go:embed licenses.txt
+var Text string